moves operator to TermPtr, where they belong and fixes issues with &= (#19)
