+++ autodiff/include/autodiff/TermPtr.h
@@ -38,8 +38,36 @@ class TermPtr
     bool operator==(const TermPtr o) const { return _ptr == o._ptr; }
     bool operator!=(const TermPtr o) const { return _ptr != o._ptr; }
 
+    TermPtr& operator&=(const TermPtr rhs);
+
   private:
     friend TermHolder;
     Term* _ptr;
 };
+
+TermPtr operator+(const TermPtr left, const TermPtr right);
+TermPtr operator*(const TermPtr left, const TermPtr right);
+TermPtr operator/(const TermPtr numerator, const TermPtr denominator);
+TermPtr operator-(const TermPtr left, const TermPtr right);
+
+TermPtr operator+(const double left, const TermPtr right);
+TermPtr operator*(const double left, const TermPtr right);
+TermPtr operator/(const double numerator, const TermPtr denominator);
+TermPtr operator-(const double left, const TermPtr right);
+
+TermPtr operator+(const TermPtr left, const double right);
+TermPtr operator*(const TermPtr left, const double right);
+TermPtr operator/(const TermPtr numerator, const double denominator);
+TermPtr operator-(const TermPtr left, const double right);
+
+TermPtr operator-(const TermPtr term);
+
+TermPtr operator!(const TermPtr term);
+TermPtr operator&(const TermPtr left, const TermPtr right);
+TermPtr operator|(const TermPtr left, const TermPtr right);
+
+TermPtr operator>(const TermPtr left, const TermPtr right);
+TermPtr operator<(const TermPtr left, const TermPtr right);
+TermPtr operator<=(const TermPtr left, const TermPtr right);
+TermPtr operator>=(const TermPtr left, const TermPtr right);
 }
